incremental

+++ src/Acqiris.cpp
@@ -11,7 +11,7 @@ using namespace CookieBox_pkg;
 
 namespace CookieBox_pkg 
 {
-	Acqiris::Acqiris(std::vector<unsigned>& lims_in , std::vector<unsigned>& baselims_in)
+	Acqiris::Acqiris(std::vector<size_t>& lims_in , std::vector<size_t>& baselims_in)
 	: m_getConfig(true)
 	, m_use(false)
 	, m_print(false)
@@ -110,16 +110,16 @@ namespace CookieBox_pkg
 	}
 
 
-	bool Acqiris::init( std::vector<unsigned>& lims_in
-			, std::vector<unsigned>& baselims_in)
+	bool Acqiris::init( std::vector<size_t>& lims_in
+			, std::vector<size_t>& baselims_in)
 	{
 		std::cerr << "lims_in.size() = " << lims_in.size() << " baselims_in.size() " << baselims_in.size() << std::endl;
 		if (! ( lims_in.size() == baselims_in.size() &&  lims_in.size() == 2 ) ){
 			std::cerr << "Failed sizes in Acqiris::init() method " << std::endl;
 			return false;
 		}
-		//std::vector<unsigned>::iterator itr = lims_in.begin();
-		//std::vector<unsigned>::iterator itr_base = baselims_in.begin();
+		//std::vector<size_t>::iterator itr = lims_in.begin();
+		//std::vector<size_t>::iterator itr_base = baselims_in.begin();
 		m_lims = lims_in;
 		m_baselims = baselims_in;
 
@@ -140,15 +140,15 @@ namespace CookieBox_pkg
 		m_srcStr = srcStr_in;
 	}
 	
-	void Acqiris::evtput(Event& evt, Env& env, const unsigned chan)
+	void Acqiris::evtput(Event& evt, Env& env, const size_t chan)
 	{
 		// use something like a 2d ndarray to pass including the chan bins to python
 		std::string outkey("aq_");
 		outkey += boost::lexical_cast<std::string>(chan);
-		unsigned shape[] = {m_data[chan].size()};
+		size_t shape[] = {m_data[chan].size()};
 		boost::shared_ptr< ndarray<double,1> > outmatPtr = boost::make_shared < ndarray<double,1> > ( shape );
 
-		for (unsigned j=0;j<m_data[chan].size();++j){
+		for (size_t j=0;j<m_data[chan].size();++j){
 			(*outmatPtr)[j] = m_data[chan][j];
 		}
 		evt.put(outmatPtr,outkey);
@@ -159,7 +159,7 @@ namespace CookieBox_pkg
 
 		std::string outkey("aq_allchans");
 
-		unsigned shape[] = {m_data.size(), m_data[0].size()};
+		size_t shape[] = {m_data.size(), m_data[0].size()};
 		boost::shared_ptr< ndarray<double,2> > outmatPtr = boost::make_shared < ndarray<double,2> > ( shape );
 
 		for (unsigned i = 0; i < m_data.size();++i){ 
@@ -302,7 +302,7 @@ namespace CookieBox_pkg
 			// Debugging the indices and the clock start time //
 			// here the 0th index can correspond to a time form 0 back to the increment size (.5ns) here 
 			// This means we could sub-divide the waveforms for computing the energy simply by adjusting by this offset.
-	   //const unsigned shape[] = {nbrChannels, nbrSamples};
+	   //const size_t shape[] = {nbrChannels, nbrSamples};
 	   //    ndarray<wform_t, 2> wf = make_ndarray<wform_t>(nbrChannels, nbrSamples);
 	   //        ndarray<wtime_t, 2> wt = make_ndarray<wtime_t>(nbrChannels, nbrSamples);
 	   //
@@ -412,7 +412,7 @@ namespace CookieBox_pkg
 		if (m_outfile.is_open())
 			m_outfile.close();
 	}
-	bool Acqiris::print_out(const unsigned eventnum)
+	bool Acqiris::print_out(const size_t eventnum)
 	{
 		std::string eventfilename(m_filename);
 		eventfilename += ".event_" + boost::lexical_cast<std::string>(eventnum);
@@ -423,7 +423,7 @@ namespace CookieBox_pkg
 			return false;
 		for (unsigned s=0;s<m_data.at(0).size();++s){
 			for (unsigned c=0;c<m_data.size();++c){
-				int val = m_data.at(c).at(s);
+				long int val = m_data.at(c).at(s);
 				m_outfile << val << "\t";
 			}
 			m_outfile << "\n";
